This fixes a problem of ACL inheritance
The root cause is permissions of individual users and groups were not
inherited when a new Attribute is created from EntityAttr. Previous
implementation only inherit permissoins of operation user and groups
that user is belonged to. This commit fixes this problem by inherit
permissions of all users and groups (and also refactored some
inefficient implementation).
(c.f. #28)