Fixed a problem that date value won't be shown at advanced search result
This fixes the problem that a value which is registered in a date typed
attribute won't be shown at the result page of advanced search. That was
because the "make_search_results" method that retrieves data which is
registered in an elasticsearch returnes string typed value for date
typed one. But template expects this value will be date type.

This removes "date" filter to align with the type which is returned by
'make_search_results". And declares Elasticsearch version that we
confirmed to work well at README.