Fixed problem to get AttributeValue which is associated with deleted EntityAttr
This commit fixes a problem that Entry.get_attrv may return
AttributeValue of Attribute which is associated with deleted EntityAttr.

In the previous implementation, get_attrv returns AttributeValue of
active Attribute. But, once EntityAttr would be deleted (technically
is_active flag is down) by editing an Entity, Attribute instances which
is associated with it wouldn't be deleted immediately. That would be
reflected until a changed Entity's Entry is accessed. And deleting an
Attribute of active EntityAttr is also operationally unexpected.

Therefore, this changed condition to get AttributeValues through
get_attrv method to return AttributeValue that refers self entry and
also refers active EntityAttr which is specified by name argument.