Refactored authentication API handler for HTTP basic authentication
In the previous implementation sends query twice to the database for
authenticating users in the AccessTokenURI. But it's enough to send it
at once. And there was no test for authentication using HTTP basic
authentication.

This commit refactores custom authentication implementation to reduce
sending unnecessary db-query and adds test for HTTP basic
authentication.