Fixed a problem in the processing of import entry
This commit fixed a proelem that occurs an Exception when an import job
is running at the situaiton that there is an abnormal Entry (e.g. there
are multiple Attributes of same name). And changed import processing to
finish job safely by making exception handler. Once an unexpected
problem might be happened in the import processing, this finishs its job
safely and recoreds it that it's an errored job.