Refactored APIv2 and fixed EntryList
This commit fixed EntryList page to list entries as alphabetical order.
Along with this fixing, this commit also refactored Entity and Entry v2
API not to send redundant requests for getting Entries that are related
with specified Entity. This integrates getEntity and getEntries API
handler for returning them with Entity information at the same time. By
this change, UI processing don't have to send multiple requests to get
Entity and Entires that are related with it.