Added processing to update es-docuemnt of referred Entries
When an Entry(e0) is updated, e0.registered() is called. But, it also needs
to call its method for each Entries that the "e0" refers to. Because each
es-documents of Entry has referrals value and they might be updated whenever
referred Entry is updated.

This commit expaned Entry.register_es() to also update es-documents of Entries
that target Entry refers to.