Change super-class of AttrType from IntFlag to IntEnum
It's impossible to read whole AttrType values by following processing as below.
```
>>> from airone.lib.types import AttrType
>>> [k.name for k in AttrType]
['OBJECT', 'STRING', 'TEXT', 'BOOLEAN', 'GROUP', 'DATE', 'ROLE', '_ARRAY', '_NAMED']
>>>
```

This is because IntFlag can't handles value that is composited one of multile flags like NAMED_OBJECT.
The AttrType is supposed to those composite values by multiple flags, so this commit changes its super
class to IntEnum to hanedle those values.